Turn vds.cpp into C-functions implementations only
If we want in future to introduce c++ tests, it would be nice to be able
to test c++ functions directly, without C wrapper.

Hence extracting everything that is not implementation of vds.h
functions into separate files, creating an internal cpp API.
